ON-17545 # Changed submission events to form events. Added approvalSteps to form
